update: student_quiz_work time remaining count from duration & expire time

+++ database/migrations/2024_05_24_201137_create_student_quizzes_table.php
@@ -20,7 +20,7 @@ return new class extends Migration {
             $table->dateTime('end_time')->nullable();
             $table->integer('duration')->unsigned()->nullable();
             $table->boolean('is_done')->nullable()->default(false);
-            $table->integer('score')->unsigned()->nullable()->default(0);
+            $table->bigInteger('score')->unsigned()->nullable()->default(0);
             $table->timestamps();
         });
     }
